add tests for conn_and_client
